extracts locations to an assests file from component small changes to language files
