+++ src/components/custom/CustomSelectInput.jsx
@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react'
 import 'react-responsive-select/dist/react-responsive-select.css'
+import locations from '../../assets/Locations'
 import i18n from '../../i18n'
 import {
   Select,
@@ -68,65 +69,4 @@ const CustomSelectInput = ({ locationValue, setInputValue, disabled }) => {
 
 export default CustomSelectInput
 
-const locations = [
-  {
-    value: 'null',
-    text: 'Any',
-    markup: <MultiSelectOptionMarkup text='Any' />,
-  },
-  {
-    value: 'Lindholmen',
-    text: 'Lindholmen',
-    markup: <MultiSelectOptionMarkup text='Lindholmen' />,
-  },
-  {
-    value: 'Gamlestaden',
-    text: 'Gamlestaden',
-    markup: <MultiSelectOptionMarkup text='Gamlestaden' />,
-  },
-  {
-    value: 'Järntorget',
-    text: 'Järntorget',
-    markup: <MultiSelectOptionMarkup text='Järntorget' />,
-  },
-  {
-    value: 'Linnè',
-    text: 'Linnè',
-    markup: <MultiSelectOptionMarkup text='Linnè' />,
-  },
-  {
-    value: 'Nordstan',
-    text: 'Nordstan',
-    markup: <MultiSelectOptionMarkup text='Nordstan' />,
-  },
-  {
-    value: 'Korsvägen',
-    text: 'Korsvägen',
-    markup: <MultiSelectOptionMarkup text='Korsvägen' />,
-  },
-  {
-    value: 'Majorna',
-    text: 'Majorna',
-    markup: <MultiSelectOptionMarkup text='Majorna' />,
-  },
-  {
-    value: 'Avenyn',
-    text: 'Avenyn',
-    markup: <MultiSelectOptionMarkup text='Avenyn' />,
-  },
-  {
-    value: 'Haga',
-    text: 'Haga',
-    markup: <MultiSelectOptionMarkup text='Haga' />,
-  },
-  {
-    value: 'Stenpiren',
-    text: 'Stenpiren',
-    markup: <MultiSelectOptionMarkup text='Stenpiren' />,
-  },
-  {
-    value: 'Eriksberg',
-    text: 'Eriksberg',
-    markup: <MultiSelectOptionMarkup text='Eriksberg' />,
-  },
-]
+
