+++ src/locales/en.js
@@ -20,16 +20,16 @@ const en = {
       location: 'Where would you like to be situated?',
     },
     answer: {
-      email: 'My email is... ',
-      size: 'We are a group of... ',
-      officeType: 'We need an... ',
+      email: 'My email is ',
+      size: 'We are a group of ',
+      officeType: 'We need an ',
       officeLable: 'Office',
       openSpaceLable: 'Open space',
-      name: 'It is called...',
+      name: 'It is called',
       companyPlaceholder: 'MyCoolCompany',
       peersNegative: 'No',
       peersPositive: 'Yes',
-      location: "We're interested in...",
+      location: "We're interested in",
       locationAll: 'Any',
     },
     submitButton: 'Submit',
